Add styles and script to treat info from github api
In this version I treated pushevents, forkevents and watchevents.
This covers most of the important events I want to show on my webpage.
New styles for the github activity display.
New area on index.html with specific ids to be used to get elements on
the new script.
This new script uses github api and adds retrieved info to the html page.